% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{ast}
\alias{ast}
\alias{ast_new}
\alias{ast_id}
\alias{ast_size}
\alias{ast_root_get}
\alias{ast_root_rep}
\alias{ast_root_set}
\title{AST APIs}
\usage{
ast_new(capacity = 10000L)

ast_id(ast)

ast_size(ast)

ast_root_get(ast)

ast_root_rep(ast, root)

ast_root_set(ast, root)
}
\arguments{
\item{capacity}{expected size of AST; size will increase dynamically if more nodes are added}

\item{ast}{AST object}

\item{root}{Root node}
}
\description{
\itemize{
\item \code{ast_new()} instantiates a new AST
\item \code{ast_id()} returns the unique id of the AST
\item \code{ast_size()} returns the number of nodes in the AST
\item \code{ast_root_get()} returns the root node of \verb{_NODE_}.
\item \code{ast_root_set()} sets the new root and deletes the old root.
\item \code{ast_root_rep()} sets the new root and returns the old root.
}
}
