% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{to_sexp}
\alias{to_df}
\alias{to_list}
\alias{to_dot}
\title{Convert AST to SEXP, data frame, list and Graphviz Dot format}
\usage{
to_sexp(ast, node = ast_root_get(ast))

to_df(ast, node = ast_root_get(ast))

to_list(ast, node = ast_root_get(ast))

to_dot(ast, node = ast_root_get(ast))
}
\arguments{
\item{ast}{AST object}

\item{node}{Node to be exported; defaults to AST root node}
}
\description{
\itemize{
\item \code{to_sexp()} turns the AST to R's AST representation.
\item \code{to_df()} turns the AST to a data frame.
\item \code{to_list()} turns the AST to an R list.
\item \code{to_dot()} turns the AST to \href{https://graphviz.org/}{Graphviz's} \href{https://graphviz.org/doc/info/lang.html}{dot} format.
}
}
