% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_aexpr}
\alias{doc_aexpr}
\alias{aexpr_new}
\alias{aexpr_ann_get}
\alias{aexpr_expr_get}
\title{Create and query \code{aexpr} nodes}
\usage{
aexpr_new(ast, ann, expr)

aexpr_ann_get(ast, node)

aexpr_expr_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{aexpr} node.}
}
\description{
\itemize{
\item \code{aexpr_ann_get()} returns the \code{ann} node of \code{aexpr}
}

\itemize{
\item \code{aexpr_expr_get()} returns the \code{expr} node of \code{aexpr}
}
}
