% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_args}
\alias{doc_args}
\alias{args_new}
\alias{args_len_get}
\alias{args_seq_get}
\alias{args_get}
\title{Create and query \code{args} nodes}
\usage{
args_new(ast, seq)

args_len_get(ast, node)

args_seq_get(ast, node)

args_get(ast, node, index)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{args} node.}

\item{`index`}{is one-based index of the child node}
}
\description{
\itemize{
\item \code{args_len_get()} returns the \code{len} node of \code{args}
}

\itemize{
\item \code{args_seq_get()} returns the child nodes as a list
}

\itemize{
\item \code{args_get()} returns the child node at position \code{index}
}
}
