% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_beg}
\alias{doc_beg}
\alias{beg_type}
\alias{beg_bom_get}
\alias{beg_loc_get}
\title{Create and query \code{beg} nodes}
\usage{
beg_type(ast, node)

beg_bom_get(ast, node)

beg_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{beg} node}
}
\description{
\itemize{
\item \code{beg_new()} instantiates a new \code{beg} node.
\item \code{beg_type()} checks if the node is a \code{beg} node.
\item \verb{beg_<child>_get()} returns the \verb{<child>} node of \code{beg}.
\item \verb{beg_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{beg_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
