% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_binop}
\alias{doc_binop}
\alias{binop_new}
\alias{binop_lexpr_get}
\alias{binop_op_get}
\alias{binop_rexpr_get}
\title{Create and query \code{binop} nodes}
\usage{
binop_new(ast, lexpr, op, rexpr)

binop_lexpr_get(ast, node)

binop_op_get(ast, node)

binop_rexpr_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{binop} node.}
}
\description{
\itemize{
\item \code{binop_lexpr_get()} returns the \code{lexpr} node of \code{binop}
}

\itemize{
\item \code{binop_op_get()} returns the \code{op} node of \code{binop}
}

\itemize{
\item \code{binop_rexpr_get()} returns the \code{rexpr} node of \code{binop}
}
}
