% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_chr}
\alias{doc_chr}
\alias{chr_type}
\alias{chr_syn_get}
\alias{chr_val_get}
\alias{chr_gap_get}
\alias{chr_loc_get}
\title{Create and query \code{chr} nodes}
\usage{
chr_type(ast, node)

chr_syn_get(ast, node)

chr_val_get(ast, node)

chr_gap_get(ast, node)

chr_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{chr} node.}
}
\description{
\itemize{
\item \code{chr_syn_get()} returns the \code{syn} node of \code{chr}
}

\itemize{
\item \code{chr_val_get()} returns the \code{val} node of \code{chr}
}

\itemize{
\item \code{chr_gap_get()} returns the \code{gap} node of \code{chr}
}

\itemize{
\item \code{chr_loc_get()} returns the \code{loc} node of \code{chr}
}
}
