% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_cond}
\alias{doc_cond}
\alias{cond_new}
\alias{cond_lbkt_get}
\alias{cond_expr_get}
\alias{cond_rbkt_get}
\title{Create and query \code{cond} nodes}
\usage{
cond_new(ast, lbkt, expr, rbkt)

cond_lbkt_get(ast, node)

cond_expr_get(ast, node)

cond_rbkt_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{cond} node.}
}
\description{
\itemize{
\item \code{cond_lbkt_get()} returns the \code{lbkt} node of \code{cond}
}

\itemize{
\item \code{cond_expr_get()} returns the \code{expr} node of \code{cond}
}

\itemize{
\item \code{cond_rbkt_get()} returns the \code{rbkt} node of \code{cond}
}
}
