% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_cpx}
\alias{doc_cpx}
\alias{cpx_type}
\alias{cpx_syn_get}
\alias{cpx_val_get}
\alias{cpx_gap_get}
\alias{cpx_gap_rep}
\alias{cpx_gap_set}
\alias{cpx_loc_get}
\title{Create and query \code{cpx} nodes}
\usage{
cpx_type(ast, node)

cpx_syn_get(ast, node)

cpx_val_get(ast, node)

cpx_gap_get(ast, node)

cpx_gap_rep(ast, node, gap)

cpx_gap_set(ast, node, gap)

cpx_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{cpx} node}

\item{gap}{\code{gap} node}
}
\description{
\itemize{
\item \code{cpx_new()} instantiates a new \code{cpx} node.
\item \code{cpx_type()} checks if the node is a \code{cpx} node.
\item \verb{cpx_<child>_get()} returns the \verb{<child>} node of \code{cpx}.
\item \verb{cpx_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{cpx_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
