% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_dbl}
\alias{doc_dbl}
\alias{dbl_type}
\alias{dbl_syn_get}
\alias{dbl_val_get}
\alias{dbl_gap_get}
\alias{dbl_gap_rep}
\alias{dbl_gap_set}
\alias{dbl_loc_get}
\title{Create and query \code{dbl} nodes}
\usage{
dbl_type(ast, node)

dbl_syn_get(ast, node)

dbl_val_get(ast, node)

dbl_gap_get(ast, node)

dbl_gap_rep(ast, node, gap)

dbl_gap_set(ast, node, gap)

dbl_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{dbl} node}

\item{gap}{\code{gap} node}
}
\description{
\itemize{
\item \code{dbl_new()} instantiates a new \code{dbl} node.
\item \code{dbl_type()} checks if the node is a \code{dbl} node.
\item \verb{dbl_<child>_get()} returns the \verb{<child>} node of \code{dbl}.
\item \verb{dbl_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{dbl_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
