% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_dlmtr}
\alias{doc_dlmtr}
\alias{dlmtr_type}
\alias{dlmtr_gap_get}
\alias{dlmtr_loc_get}
\title{Create and query \code{dlmtr} nodes}
\usage{
dlmtr_type(ast, node)

dlmtr_gap_get(ast, node)

dlmtr_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{dlmtr} node.}
}
\description{
\itemize{
\item \code{dlmtr_gap_get()} returns the \code{gap} node of \code{dlmtr}
}

\itemize{
\item \code{dlmtr_loc_get()} returns the \code{loc} node of \code{dlmtr}
}
}
