% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_dpar}
\alias{doc_dpar}
\alias{dpar_new}
\alias{dpar_name_get}
\alias{dpar_op_get}
\alias{dpar_expr_get}
\title{Create and query \code{dpar} nodes}
\usage{
dpar_new(ast, name, op, expr)

dpar_name_get(ast, node)

dpar_op_get(ast, node)

dpar_expr_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{dpar} node.}
}
\description{
\itemize{
\item \code{dpar_name_get()} returns the \code{name} node of \code{dpar}
}

\itemize{
\item \code{dpar_op_get()} returns the \code{op} node of \code{dpar}
}

\itemize{
\item \code{dpar_expr_get()} returns the \code{expr} node of \code{dpar}
}
}
