% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_end}
\alias{doc_end}
\alias{end_new}
\alias{end_gap_get}
\alias{end_loc_get}
\title{Create and query \code{end} nodes}
\usage{
end_new(ast, gap, loc)

end_gap_get(ast, node)

end_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{end} node.}
}
\description{
\itemize{
\item \code{end_gap_get()} returns the \code{gap} node of \code{end}
}

\itemize{
\item \code{end_loc_get()} returns the \code{loc} node of \code{end}
}
}
