% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_err}
\alias{doc_err}
\alias{err_type}
\alias{err_msg_get}
\alias{err_loc_get}
\title{Create and query \code{err} nodes}
\usage{
err_type(ast, node)

err_msg_get(ast, node)

err_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{err} node.}
}
\description{
\itemize{
\item \code{err_msg_get()} returns the \code{msg} node of \code{err}
}

\itemize{
\item \code{err_loc_get()} returns the \code{loc} node of \code{err}
}
}
