% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_exprs}
\alias{doc_exprs}
\alias{exprs_new}
\alias{exprs_len_get}
\alias{exprs_seq_get}
\alias{exprs_get}
\title{Create and query \code{exprs} nodes}
\usage{
exprs_new(ast, seq)

exprs_len_get(ast, node)

exprs_seq_get(ast, node)

exprs_get(ast, node, index)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{exprs} node.}

\item{`index`}{is one-based index of the child node}
}
\description{
\itemize{
\item \code{exprs_len_get()} returns the \code{len} node of \code{exprs}
}

\itemize{
\item \code{exprs_seq_get()} returns the child nodes as a list
}

\itemize{
\item \code{exprs_get()} returns the child node at position \code{index}
}
}
