% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_fncall}
\alias{doc_fncall}
\alias{fncall_new}
\alias{fncall_fn_get}
\alias{fncall_lbkt_get}
\alias{fncall_args_get}
\alias{fncall_rbkt_get}
\title{Create and query \code{fncall} nodes}
\usage{
fncall_new(ast, fn, lbkt, args, rbkt)

fncall_fn_get(ast, node)

fncall_lbkt_get(ast, node)

fncall_args_get(ast, node)

fncall_rbkt_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{fncall} node.}
}
\description{
\itemize{
\item \code{fncall_fn_get()} returns the \code{fn} node of \code{fncall}
}

\itemize{
\item \code{fncall_lbkt_get()} returns the \code{lbkt} node of \code{fncall}
}

\itemize{
\item \code{fncall_args_get()} returns the \code{args} node of \code{fncall}
}

\itemize{
\item \code{fncall_rbkt_get()} returns the \code{rbkt} node of \code{fncall}
}
}
