% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_gap}
\alias{doc_gap}
\alias{gap_type}
\alias{gap_val_get}
\alias{gap_loc_get}
\title{Create and query \code{gap} nodes}
\usage{
gap_type(ast, node)

gap_val_get(ast, node)

gap_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{gap} node.}
}
\description{
\itemize{
\item \code{gap_val_get()} returns the \code{val} node of \code{gap}
}

\itemize{
\item \code{gap_loc_get()} returns the \code{loc} node of \code{gap}
}
}
