% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_icond}
\alias{doc_icond}
\alias{icond_new}
\alias{icond_iop_get}
\alias{icond_cond_get}
\alias{icond_ibody_get}
\title{Create and query \code{icond} nodes}
\usage{
icond_new(ast, iop, cond, ibody)

icond_iop_get(ast, node)

icond_cond_get(ast, node)

icond_ibody_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{icond} node.}
}
\description{
\itemize{
\item \code{icond_iop_get()} returns the \code{iop} node of \code{icond}
}

\itemize{
\item \code{icond_cond_get()} returns the \code{cond} node of \code{icond}
}

\itemize{
\item \code{icond_ibody_get()} returns the \code{ibody} node of \code{icond}
}
}
