% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_iecond}
\alias{doc_iecond}
\alias{iecond_new}
\alias{iecond_iop_get}
\alias{iecond_cond_get}
\alias{iecond_ibody_get}
\alias{iecond_eop_get}
\alias{iecond_ebody_get}
\title{Create and query \code{iecond} nodes}
\usage{
iecond_new(ast, iop, cond, ibody, eop, ebody)

iecond_iop_get(ast, node)

iecond_cond_get(ast, node)

iecond_ibody_get(ast, node)

iecond_eop_get(ast, node)

iecond_ebody_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{iecond} node.}
}
\description{
\itemize{
\item \code{iecond_iop_get()} returns the \code{iop} node of \code{iecond}
}

\itemize{
\item \code{iecond_cond_get()} returns the \code{cond} node of \code{iecond}
}

\itemize{
\item \code{iecond_ibody_get()} returns the \code{ibody} node of \code{iecond}
}

\itemize{
\item \code{iecond_eop_get()} returns the \code{eop} node of \code{iecond}
}

\itemize{
\item \code{iecond_ebody_get()} returns the \code{ebody} node of \code{iecond}
}
}
