% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_int}
\alias{doc_int}
\alias{int_type}
\alias{int_syn_get}
\alias{int_val_get}
\alias{int_gap_get}
\alias{int_loc_get}
\title{Create and query \code{int} nodes}
\usage{
int_type(ast, node)

int_syn_get(ast, node)

int_val_get(ast, node)

int_gap_get(ast, node)

int_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{int} node.}
}
\description{
\itemize{
\item \code{int_syn_get()} returns the \code{syn} node of \code{int}
}

\itemize{
\item \code{int_val_get()} returns the \code{val} node of \code{int}
}

\itemize{
\item \code{int_gap_get()} returns the \code{gap} node of \code{int}
}

\itemize{
\item \code{int_loc_get()} returns the \code{loc} node of \code{int}
}
}
