% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_iter}
\alias{doc_iter}
\alias{iter_new}
\alias{iter_lbkt_get}
\alias{iter_var_get}
\alias{iter_op_get}
\alias{iter_expr_get}
\alias{iter_rbkt_get}
\title{Create and query \code{iter} nodes}
\usage{
iter_new(ast, lbkt, var, op, expr, rbkt)

iter_lbkt_get(ast, node)

iter_var_get(ast, node)

iter_op_get(ast, node)

iter_expr_get(ast, node)

iter_rbkt_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{iter} node.}
}
\description{
\itemize{
\item \code{iter_lbkt_get()} returns the \code{lbkt} node of \code{iter}
}

\itemize{
\item \code{iter_var_get()} returns the \code{var} node of \code{iter}
}

\itemize{
\item \code{iter_op_get()} returns the \code{op} node of \code{iter}
}

\itemize{
\item \code{iter_expr_get()} returns the \code{expr} node of \code{iter}
}

\itemize{
\item \code{iter_rbkt_get()} returns the \code{rbkt} node of \code{iter}
}
}
