% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_lgl}
\alias{doc_lgl}
\alias{lgl_type}
\alias{lgl_val_get}
\alias{lgl_gap_get}
\alias{lgl_gap_rep}
\alias{lgl_gap_set}
\alias{lgl_loc_get}
\title{Create and query \code{lgl} nodes}
\usage{
lgl_type(ast, node)

lgl_val_get(ast, node)

lgl_gap_get(ast, node)

lgl_gap_rep(ast, node, gap)

lgl_gap_set(ast, node, gap)

lgl_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{lgl} node}

\item{gap}{\code{gap} node}
}
\description{
\itemize{
\item \code{lgl_new()} instantiates a new \code{lgl} node.
\item \code{lgl_type()} checks if the node is a \code{lgl} node.
\item \verb{lgl_<child>_get()} returns the \verb{<child>} node of \code{lgl}.
\item \verb{lgl_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{lgl_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
