% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_loc}
\alias{doc_loc}
\alias{loc_type}
\alias{loc_lrow_get}
\alias{loc_lcol_get}
\alias{loc_lchr_get}
\alias{loc_lbyte_get}
\alias{loc_rrow_get}
\alias{loc_rcol_get}
\alias{loc_rchr_get}
\alias{loc_rbyte_get}
\title{Create and query \code{loc} nodes}
\usage{
loc_type(ast, node)

loc_lrow_get(ast, node)

loc_lcol_get(ast, node)

loc_lchr_get(ast, node)

loc_lbyte_get(ast, node)

loc_rrow_get(ast, node)

loc_rcol_get(ast, node)

loc_rchr_get(ast, node)

loc_rbyte_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{loc} node}
}
\description{
\itemize{
\item \code{loc_new()} instantiates a new \code{loc} node.
\item \code{loc_type()} checks if the node is a \code{loc} node.
\item \verb{loc_<child>_get()} returns the \verb{<child>} node of \code{loc}.
\item \verb{loc_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{loc_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
