% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_msng}
\alias{doc_msng}
\alias{msng_new}
\alias{msng_gap_get}
\alias{msng_loc_get}
\title{Create and query \code{msng} nodes}
\usage{
msng_new(ast, gap, loc)

msng_gap_get(ast, node)

msng_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{msng} node.}
}
\description{
\itemize{
\item \code{msng_gap_get()} returns the \code{gap} node of \code{msng}
}

\itemize{
\item \code{msng_loc_get()} returns the \code{loc} node of \code{msng}
}
}
