% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_narg}
\alias{doc_narg}
\alias{narg_new}
\alias{narg_name_get}
\alias{narg_op_get}
\alias{narg_expr_get}
\title{Create and query \code{narg} nodes}
\usage{
narg_new(ast, name, op, expr)

narg_name_get(ast, node)

narg_op_get(ast, node)

narg_expr_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{narg} node.}
}
\description{
\itemize{
\item \code{narg_name_get()} returns the \code{name} node of \code{narg}
}

\itemize{
\item \code{narg_op_get()} returns the \code{op} node of \code{narg}
}

\itemize{
\item \code{narg_expr_get()} returns the \code{expr} node of \code{narg}
}
}
