% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_null}
\alias{doc_null}
\alias{null_new}
\alias{null_gap_get}
\alias{null_loc_get}
\title{Create and query \code{null} nodes}
\usage{
null_new(ast, gap, loc)

null_gap_get(ast, node)

null_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{null} node.}
}
\description{
\itemize{
\item \code{null_gap_get()} returns the \code{gap} node of \code{null}
}

\itemize{
\item \code{null_loc_get()} returns the \code{loc} node of \code{null}
}
}
