% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_pars}
\alias{doc_pars}
\alias{pars_new}
\alias{pars_len_get}
\alias{pars_seq_get}
\alias{pars_get}
\title{Create and query \code{pars} nodes}
\usage{
pars_new(ast, seq)

pars_len_get(ast, node)

pars_seq_get(ast, node)

pars_get(ast, node, index)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{pars} node.}

\item{`index`}{is one-based index of the child node}
}
\description{
\itemize{
\item \code{pars_len_get()} returns the \code{len} node of \code{pars}
}

\itemize{
\item \code{pars_seq_get()} returns the child nodes as a list
}

\itemize{
\item \code{pars_get()} returns the child node at position \code{index}
}
}
