% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_pgm}
\alias{doc_pgm}
\alias{pgm_new}
\alias{pgm_beg_get}
\alias{pgm_exprs_get}
\alias{pgm_end_get}
\title{Create and query \code{pgm} nodes}
\usage{
pgm_new(ast, beg, exprs, end)

pgm_beg_get(ast, node)

pgm_exprs_get(ast, node)

pgm_end_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{pgm} node.}
}
\description{
\itemize{
\item \code{pgm_beg_get()} returns the \code{beg} node of \code{pgm}
}

\itemize{
\item \code{pgm_exprs_get()} returns the \code{exprs} node of \code{pgm}
}

\itemize{
\item \code{pgm_end_get()} returns the \code{end} node of \code{pgm}
}
}
