% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_rlp}
\alias{doc_rlp}
\alias{rlp_new}
\alias{rlp_op_get}
\alias{rlp_body_get}
\title{Create and query \code{rlp} nodes}
\usage{
rlp_new(ast, op, body)

rlp_op_get(ast, node)

rlp_body_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{rlp} node.}
}
\description{
\itemize{
\item \code{rlp_op_get()} returns the \code{op} node of \code{rlp}
}

\itemize{
\item \code{rlp_body_get()} returns the \code{body} node of \code{rlp}
}
}
