% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_sym}
\alias{doc_sym}
\alias{sym_type}
\alias{sym_syn_get}
\alias{sym_val_get}
\alias{sym_gap_get}
\alias{sym_gap_rep}
\alias{sym_gap_set}
\alias{sym_loc_get}
\title{Create and query \code{sym} nodes}
\usage{
sym_type(ast, node)

sym_syn_get(ast, node)

sym_val_get(ast, node)

sym_gap_get(ast, node)

sym_gap_rep(ast, node, gap)

sym_gap_set(ast, node, gap)

sym_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{sym} node}

\item{gap}{\code{gap} node}
}
\description{
\itemize{
\item \code{sym_new()} instantiates a new \code{sym} node.
\item \code{sym_type()} checks if the node is a \code{sym} node.
\item \verb{sym_<child>_get()} returns the \verb{<child>} node of \code{sym}.
\item \verb{sym_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{sym_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
