% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\alias{parse_str}
\alias{parse_file}
\alias{unparse_str}
\alias{unparse_file}
\title{Parse and Unparse AST}
\usage{
parse_str(str)

parse_file(file)

unparse_str(ast, node = ast_root_get(ast))

unparse_file(ast, node = ast_root_get(ast), file)
}
\arguments{
\item{str}{A scalar string to be parsed}

\item{file}{File to be parsed}

\item{ast}{AST object}

\item{node}{Node to be unparsed; defaults to AST root node}
}
\description{
\itemize{
\item \code{parse_str()} parses AST from an R string
\item \code{parse_file()} parses AST from a file
\item \code{unparse_str()} unparses AST to an R string
\item \code{unparse_file()} unparses AST to a file
}
}
