% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugr.R
\name{sugr}
\alias{sugr}
\alias{desugar}
\title{Desugar AST}
\usage{
desugar(ast, node = ast_root_get(ast), strictness = TRUE)
}
\arguments{
\item{ast}{AST object}

\item{node}{starting node for desugaring; defaults to AST root}
}
\description{
\itemize{
\item \code{desugar()} Desugars AST in-place to well-formed R syntax
}
}
