% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_beg}
\alias{doc_beg}
\alias{beg_type}
\alias{beg_bom_get}
\alias{beg_loc_get}
\title{Create and query \code{beg} nodes}
\usage{
beg_type(ast, node)

beg_bom_get(ast, node)

beg_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{beg} node.}
}
\description{
\itemize{
\item \code{beg_bom_get()} returns the \code{bom} node of \code{beg}
}

\itemize{
\item \code{beg_loc_get()} returns the \code{loc} node of \code{beg}
}
}
