% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_bkt}
\alias{doc_bkt}
\alias{bkt_type}
\alias{bkt_gap_get}
\alias{bkt_loc_get}
\title{Create and query \code{bkt} nodes}
\usage{
bkt_type(ast, node)

bkt_gap_get(ast, node)

bkt_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{bkt} node.}
}
\description{
\itemize{
\item \code{bkt_gap_get()} returns the \code{gap} node of \code{bkt}
}

\itemize{
\item \code{bkt_loc_get()} returns the \code{loc} node of \code{bkt}
}
}
