% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_blk}
\alias{doc_blk}
\alias{blk_new}
\alias{blk_lbkt_get}
\alias{blk_exprs_get}
\alias{blk_rbkt_get}
\title{Create and query \code{blk} nodes}
\usage{
blk_new(ast, lbkt, exprs, rbkt)

blk_lbkt_get(ast, node)

blk_exprs_get(ast, node)

blk_rbkt_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{blk} node.}
}
\description{
\itemize{
\item \code{blk_lbkt_get()} returns the \code{lbkt} node of \code{blk}
}

\itemize{
\item \code{blk_exprs_get()} returns the \code{exprs} node of \code{blk}
}

\itemize{
\item \code{blk_rbkt_get()} returns the \code{rbkt} node of \code{blk}
}
}
