% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_chr}
\alias{doc_chr}
\alias{chr_type}
\alias{chr_syn_get}
\alias{chr_val_get}
\alias{chr_gap_get}
\alias{chr_gap_rep}
\alias{chr_gap_set}
\alias{chr_loc_get}
\title{Create and query \code{chr} nodes}
\usage{
chr_type(ast, node)

chr_syn_get(ast, node)

chr_val_get(ast, node)

chr_gap_get(ast, node)

chr_gap_rep(ast, node, gap)

chr_gap_set(ast, node, gap)

chr_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{chr} node}

\item{gap}{\code{gap} node}
}
\description{
\itemize{
\item \code{chr_new()} instantiates a new \code{chr} node.
\item \code{chr_type()} checks if the node is a \code{chr} node.
\item \verb{chr_<child>_get()} returns the \verb{<child>} node of \code{chr}.
\item \verb{chr_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{chr_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
