% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_cpx}
\alias{doc_cpx}
\alias{cpx_type}
\alias{cpx_syn_get}
\alias{cpx_val_get}
\alias{cpx_gap_get}
\alias{cpx_loc_get}
\title{Create and query \code{cpx} nodes}
\usage{
cpx_type(ast, node)

cpx_syn_get(ast, node)

cpx_val_get(ast, node)

cpx_gap_get(ast, node)

cpx_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{cpx} node.}
}
\description{
\itemize{
\item \code{cpx_syn_get()} returns the \code{syn} node of \code{cpx}
}

\itemize{
\item \code{cpx_val_get()} returns the \code{val} node of \code{cpx}
}

\itemize{
\item \code{cpx_gap_get()} returns the \code{gap} node of \code{cpx}
}

\itemize{
\item \code{cpx_loc_get()} returns the \code{loc} node of \code{cpx}
}
}
