% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_dbl}
\alias{doc_dbl}
\alias{dbl_type}
\alias{dbl_syn_get}
\alias{dbl_val_get}
\alias{dbl_gap_get}
\alias{dbl_loc_get}
\title{Create and query \code{dbl} nodes}
\usage{
dbl_type(ast, node)

dbl_syn_get(ast, node)

dbl_val_get(ast, node)

dbl_gap_get(ast, node)

dbl_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{dbl} node.}
}
\description{
\itemize{
\item \code{dbl_syn_get()} returns the \code{syn} node of \code{dbl}
}

\itemize{
\item \code{dbl_val_get()} returns the \code{val} node of \code{dbl}
}

\itemize{
\item \code{dbl_gap_get()} returns the \code{gap} node of \code{dbl}
}

\itemize{
\item \code{dbl_loc_get()} returns the \code{loc} node of \code{dbl}
}
}
