% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_dlmtd}
\alias{doc_dlmtd}
\alias{dlmtd_new}
\alias{dlmtd_expr_get}
\alias{dlmtd_dlmtr_get}
\title{Create and query \code{dlmtd} nodes}
\usage{
dlmtd_new(ast, expr, dlmtr)

dlmtd_expr_get(ast, node)

dlmtd_dlmtr_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{dlmtd} node.}
}
\description{
\itemize{
\item \code{dlmtd_expr_get()} returns the \code{expr} node of \code{dlmtd}
}

\itemize{
\item \code{dlmtd_dlmtr_get()} returns the \code{dlmtr} node of \code{dlmtd}
}
}
