% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_dlmtr}
\alias{doc_dlmtr}
\alias{dlmtr_type}
\alias{dlmtr_gap_get}
\alias{dlmtr_gap_rep}
\alias{dlmtr_gap_set}
\alias{dlmtr_loc_get}
\title{Create and query \code{dlmtr} nodes}
\usage{
dlmtr_type(ast, node)

dlmtr_gap_get(ast, node)

dlmtr_gap_rep(ast, node, gap)

dlmtr_gap_set(ast, node, gap)

dlmtr_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{dlmtr} node}

\item{gap}{\code{gap} node}
}
\description{
\itemize{
\item \code{dlmtr_new()} instantiates a new \code{dlmtr} node.
\item \code{dlmtr_type()} checks if the node is a \code{dlmtr} node.
\item \verb{dlmtr_<child>_get()} returns the \verb{<child>} node of \code{dlmtr}.
\item \verb{dlmtr_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{dlmtr_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
