% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_err}
\alias{doc_err}
\alias{err_type}
\alias{err_msg_get}
\alias{err_loc_get}
\title{Create and query \code{err} nodes}
\usage{
err_type(ast, node)

err_msg_get(ast, node)

err_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{err} node}
}
\description{
\itemize{
\item \code{err_new()} instantiates a new \code{err} node.
\item \code{err_type()} checks if the node is a \code{err} node.
\item \verb{err_<child>_get()} returns the \verb{<child>} node of \code{err}.
\item \verb{err_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{err_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
