% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_flp}
\alias{doc_flp}
\alias{flp_new}
\alias{flp_op_get}
\alias{flp_iter_get}
\alias{flp_body_get}
\title{Create and query \code{flp} nodes}
\usage{
flp_new(ast, op, iter, body)

flp_op_get(ast, node)

flp_iter_get(ast, node)

flp_body_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{flp} node.}
}
\description{
\itemize{
\item \code{flp_op_get()} returns the \code{op} node of \code{flp}
}

\itemize{
\item \code{flp_iter_get()} returns the \code{iter} node of \code{flp}
}

\itemize{
\item \code{flp_body_get()} returns the \code{body} node of \code{flp}
}
}
