% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_fndefn}
\alias{doc_fndefn}
\alias{fndefn_new}
\alias{fndefn_op_get}
\alias{fndefn_lbkt_get}
\alias{fndefn_pars_get}
\alias{fndefn_rbkt_get}
\alias{fndefn_body_get}
\title{Create and query \code{fndefn} nodes}
\usage{
fndefn_new(ast, op, lbkt, pars, rbkt, body)

fndefn_op_get(ast, node)

fndefn_lbkt_get(ast, node)

fndefn_pars_get(ast, node)

fndefn_rbkt_get(ast, node)

fndefn_body_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{fndefn} node.}
}
\description{
\itemize{
\item \code{fndefn_op_get()} returns the \code{op} node of \code{fndefn}
}

\itemize{
\item \code{fndefn_lbkt_get()} returns the \code{lbkt} node of \code{fndefn}
}

\itemize{
\item \code{fndefn_pars_get()} returns the \code{pars} node of \code{fndefn}
}

\itemize{
\item \code{fndefn_rbkt_get()} returns the \code{rbkt} node of \code{fndefn}
}

\itemize{
\item \code{fndefn_body_get()} returns the \code{body} node of \code{fndefn}
}
}
