% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_gap}
\alias{doc_gap}
\alias{gap_type}
\alias{gap_val_get}
\alias{gap_loc_get}
\title{Create and query \code{gap} nodes}
\usage{
gap_type(ast, node)

gap_val_get(ast, node)

gap_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{gap} node}
}
\description{
\itemize{
\item \code{gap_new()} instantiates a new \code{gap} node.
\item \code{gap_type()} checks if the node is a \code{gap} node.
\item \verb{gap_<child>_get()} returns the \verb{<child>} node of \code{gap}.
\item \verb{gap_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{gap_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
