% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_grp}
\alias{doc_grp}
\alias{grp_new}
\alias{grp_lbkt_get}
\alias{grp_expr_get}
\alias{grp_rbkt_get}
\title{Create and query \code{grp} nodes}
\usage{
grp_new(ast, lbkt, expr, rbkt)

grp_lbkt_get(ast, node)

grp_expr_get(ast, node)

grp_rbkt_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{grp} node.}
}
\description{
\itemize{
\item \code{grp_lbkt_get()} returns the \code{lbkt} node of \code{grp}
}

\itemize{
\item \code{grp_expr_get()} returns the \code{expr} node of \code{grp}
}

\itemize{
\item \code{grp_rbkt_get()} returns the \code{rbkt} node of \code{grp}
}
}
