% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_idx}
\alias{doc_idx}
\alias{idx_new}
\alias{idx_obj_get}
\alias{idx_lbkt_get}
\alias{idx_args_get}
\alias{idx_rbkt1_get}
\alias{idx_rbkt2_get}
\title{Create and query \code{idx} nodes}
\usage{
idx_new(ast, obj, lbkt, args, rbkt1, rbkt2)

idx_obj_get(ast, node)

idx_lbkt_get(ast, node)

idx_args_get(ast, node)

idx_rbkt1_get(ast, node)

idx_rbkt2_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{idx} node.}
}
\description{
\itemize{
\item \code{idx_obj_get()} returns the \code{obj} node of \code{idx}
}

\itemize{
\item \code{idx_lbkt_get()} returns the \code{lbkt} node of \code{idx}
}

\itemize{
\item \code{idx_args_get()} returns the \code{args} node of \code{idx}
}

\itemize{
\item \code{idx_rbkt1_get()} returns the \code{rbkt1} node of \code{idx}
}

\itemize{
\item \code{idx_rbkt2_get()} returns the \code{rbkt2} node of \code{idx}
}
}
