% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_int}
\alias{doc_int}
\alias{int_type}
\alias{int_syn_get}
\alias{int_val_get}
\alias{int_gap_get}
\alias{int_gap_rep}
\alias{int_gap_set}
\alias{int_loc_get}
\title{Create and query \code{int} nodes}
\usage{
int_type(ast, node)

int_syn_get(ast, node)

int_val_get(ast, node)

int_gap_get(ast, node)

int_gap_rep(ast, node, gap)

int_gap_set(ast, node, gap)

int_loc_get(ast, node)
}
\arguments{
\item{ast}{AST object}

\item{node}{\code{int} node}

\item{gap}{\code{gap} node}
}
\description{
\itemize{
\item \code{int_new()} instantiates a new \code{int} node.
\item \code{int_type()} checks if the node is a \code{int} node.
\item \verb{int_<child>_get()} returns the \verb{<child>} node of \code{int}.
\item \verb{int_<child>_set()} sets the new \verb{<child>} and deletes the old \verb{<child>}.
\item \verb{int_<child>_rep()} sets the new \verb{<child>} and returns the old \verb{<child>}.
}
}
