% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_lgl}
\alias{doc_lgl}
\alias{lgl_type}
\alias{lgl_val_get}
\alias{lgl_gap_get}
\alias{lgl_loc_get}
\title{Create and query \code{lgl} nodes}
\usage{
lgl_type(ast, node)

lgl_val_get(ast, node)

lgl_gap_get(ast, node)

lgl_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{lgl} node.}
}
\description{
\itemize{
\item \code{lgl_val_get()} returns the \code{val} node of \code{lgl}
}

\itemize{
\item \code{lgl_gap_get()} returns the \code{gap} node of \code{lgl}
}

\itemize{
\item \code{lgl_loc_get()} returns the \code{loc} node of \code{lgl}
}
}
