% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_loc}
\alias{doc_loc}
\alias{loc_type}
\alias{loc_lrow_get}
\alias{loc_lcol_get}
\alias{loc_lchr_get}
\alias{loc_lbyte_get}
\alias{loc_rrow_get}
\alias{loc_rcol_get}
\alias{loc_rchr_get}
\alias{loc_rbyte_get}
\title{Create and query \code{loc} nodes}
\usage{
loc_type(ast, node)

loc_lrow_get(ast, node)

loc_lcol_get(ast, node)

loc_lchr_get(ast, node)

loc_lbyte_get(ast, node)

loc_rrow_get(ast, node)

loc_rcol_get(ast, node)

loc_rchr_get(ast, node)

loc_rbyte_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{loc} node.}
}
\description{
\itemize{
\item \code{loc_lrow_get()} returns the \code{lrow} node of \code{loc}
}

\itemize{
\item \code{loc_lcol_get()} returns the \code{lcol} node of \code{loc}
}

\itemize{
\item \code{loc_lchr_get()} returns the \code{lchr} node of \code{loc}
}

\itemize{
\item \code{loc_lbyte_get()} returns the \code{lbyte} node of \code{loc}
}

\itemize{
\item \code{loc_rrow_get()} returns the \code{rrow} node of \code{loc}
}

\itemize{
\item \code{loc_rcol_get()} returns the \code{rcol} node of \code{loc}
}

\itemize{
\item \code{loc_rchr_get()} returns the \code{rchr} node of \code{loc}
}

\itemize{
\item \code{loc_rbyte_get()} returns the \code{rbyte} node of \code{loc}
}
}
