% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_ndpar}
\alias{doc_ndpar}
\alias{ndpar_new}
\alias{ndpar_name_get}
\title{Create and query \code{ndpar} nodes}
\usage{
ndpar_new(ast, name)

ndpar_name_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{ndpar} node.}
}
\description{
\itemize{
\item \code{ndpar_name_get()} returns the \code{name} node of \code{ndpar}
}
}
