% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_nuop}
\alias{doc_nuop}
\alias{nuop_new}
\alias{nuop_op_get}
\title{Create and query \code{nuop} nodes}
\usage{
nuop_new(ast, op)

nuop_op_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{nuop} node.}
}
\description{
\itemize{
\item \code{nuop_op_get()} returns the \code{op} node of \code{nuop}
}
}
