% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_op}
\alias{doc_op}
\alias{op_type}
\alias{op_gap_get}
\alias{op_loc_get}
\title{Create and query \code{op} nodes}
\usage{
op_type(ast, node)

op_gap_get(ast, node)

op_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{op} node.}
}
\description{
\itemize{
\item \code{op_gap_get()} returns the \code{gap} node of \code{op}
}

\itemize{
\item \code{op_loc_get()} returns the \code{loc} node of \code{op}
}
}
