% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_parg}
\alias{doc_parg}
\alias{parg_new}
\alias{parg_expr_get}
\title{Create and query \code{parg} nodes}
\usage{
parg_new(ast, expr)

parg_expr_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{parg} node.}
}
\description{
\itemize{
\item \code{parg_expr_get()} returns the \code{expr} node of \code{parg}
}
}
