% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_sub}
\alias{doc_sub}
\alias{sub_new}
\alias{sub_obj_get}
\alias{sub_lbkt_get}
\alias{sub_args_get}
\alias{sub_rbkt_get}
\title{Create and query \code{sub} nodes}
\usage{
sub_new(ast, obj, lbkt, args, rbkt)

sub_obj_get(ast, node)

sub_lbkt_get(ast, node)

sub_args_get(ast, node)

sub_rbkt_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{sub} node.}
}
\description{
\itemize{
\item \code{sub_obj_get()} returns the \code{obj} node of \code{sub}
}

\itemize{
\item \code{sub_lbkt_get()} returns the \code{lbkt} node of \code{sub}
}

\itemize{
\item \code{sub_args_get()} returns the \code{args} node of \code{sub}
}

\itemize{
\item \code{sub_rbkt_get()} returns the \code{rbkt} node of \code{sub}
}
}
