% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_sym}
\alias{doc_sym}
\alias{sym_type}
\alias{sym_syn_get}
\alias{sym_val_get}
\alias{sym_gap_get}
\alias{sym_loc_get}
\title{Create and query \code{sym} nodes}
\usage{
sym_type(ast, node)

sym_syn_get(ast, node)

sym_val_get(ast, node)

sym_gap_get(ast, node)

sym_loc_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{sym} node.}
}
\description{
\itemize{
\item \code{sym_syn_get()} returns the \code{syn} node of \code{sym}
}

\itemize{
\item \code{sym_val_get()} returns the \code{val} node of \code{sym}
}

\itemize{
\item \code{sym_gap_get()} returns the \code{gap} node of \code{sym}
}

\itemize{
\item \code{sym_loc_get()} returns the \code{loc} node of \code{sym}
}
}
