% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_unop}
\alias{doc_unop}
\alias{unop_new}
\alias{unop_op_get}
\alias{unop_expr_get}
\title{Create and query \code{unop} nodes}
\usage{
unop_new(ast, op, expr)

unop_op_get(ast, node)

unop_expr_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{unop} node.}
}
\description{
\itemize{
\item \code{unop_op_get()} returns the \code{op} node of \code{unop}
}

\itemize{
\item \code{unop_expr_get()} returns the \code{expr} node of \code{unop}
}
}
