% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_node.R
\name{doc_wlp}
\alias{doc_wlp}
\alias{wlp_new}
\alias{wlp_op_get}
\alias{wlp_cond_get}
\alias{wlp_body_get}
\title{Create and query \code{wlp} nodes}
\usage{
wlp_new(ast, op, cond, body)

wlp_op_get(ast, node)

wlp_cond_get(ast, node)

wlp_body_get(ast, node)
}
\arguments{
\item{ast}{The AST object.}

\item{node}{The \code{wlp} node.}
}
\description{
\itemize{
\item \code{wlp_op_get()} returns the \code{op} node of \code{wlp}
}

\itemize{
\item \code{wlp_cond_get()} returns the \code{cond} node of \code{wlp}
}

\itemize{
\item \code{wlp_body_get()} returns the \code{body} node of \code{wlp}
}
}
